+++ example/android/shared/src/commonMain/kotlin/com/myrnproject/shared/RNViewManagersList.kt
@@ -1,0 +1,11 @@
+package com.myrnproject.shared
+
+import de.voize.reaktnativetoolkit.util.ReactNativeViewManagerProvider
+
+fun getReactNativeViewManagerProviders(
+    persistentConfig: PersistentConfig,
+): List<ReactNativeViewManagerProvider> {
+    return listOf(
+        MyComposeViewRNViewManagerProvider(persistentConfig),
+    )
+}