+++ kotlin/reakt-native-toolkit/src/androidMain/kotlin/de/voize/reaktnativetoolkit/util/ReactNativeViewManagerProvider.kt
@@ -1,0 +1,11 @@
+package de.voize.reaktnativetoolkit.util
+
+import com.facebook.react.uimanager.ViewManager
+
+actual interface ReactNativeViewManagerProvider {
+    fun getViewManager(): ViewManager<*, *>
+}
+
+fun List<ReactNativeViewManagerProvider>.getViewManagers(): List<ViewManager<*, *>> {
+    return map { it.getViewManager() }
+}