+++ kotlin/reakt-native-toolkit/src/commonMain/kotlin/de/voize/reaktnativetoolkit/util/ReactNativeViewManagerProvider.kt
@@ -1,0 +1,3 @@
+package de.voize.reaktnativetoolkit.util
+
+expect interface ReactNativeViewManagerProvider