+++ CHANGELOG.md
@@ -3,8 +3,10 @@
 ## unreleased
 
 - Update example app iOS part to correct setup for RN 0.74.0
+- Implement ReactNativeViewManager annotation to render Compose Multiplatform components in React Native
 
 ## v0.17.0
+
 ## v0.16.2
 
 - Build toolkit from source in example project by default